Add files via upload
This is a web scraper that I created using Python. The program scrapes the name and current price of a cryptocurrency, converts them to string values, and adds them to a data frame. The data frame is then written to a csv file. If the csv file does not already exist, it creates it. If it does already exist, it appends the latest values to the file. I used simple automation to automate the function to run once per day.