Add files via upload
Using Python, this Web Scraper pulls text from a website (Martin Luther King Jr's I Have A Dream speech), cleans the text of all punctuation, converts it to a list, and then obtains value counts for each word in the speech. It then saves the result to a data frame and saves it to a CSV file.