Add files via upload
Very similar to the automated crypto puller project that I've also uploaded to this folder, this Kaspa web scraper collects the name and price of Kaspa, converts it to text format, timestamps when the data collection occurs, and then adds these values to a data frame, which has then been converted to a CSV file stored on my computer. I again used the 'time.sleep()' function to automate the process.